+++ src/main/java/com/gongzone/central/member/login/domain/LoginRequest.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class LoginRequest {
-    private String loginId;   // 아이디
-    private String loginPw;     // 비번
+    private String loginId;
+    private String loginPw;
     private String userAgent;
 }
