+++ src/main/java/com/gongzone/central/member/login/mapper/LoginMapper.java
@@ -9,7 +9,6 @@ public interface LoginMapper {
     Member getMemberById(String loginId);
     Member getMemberByEmail(String email);
     Member getMemberByNo(String memberNo);
-    //String getPointNoByMemberNo(String memberNo);
 
     void loginLogInsert(LoginLog loginLog);
     int logoutLogUpdate(int loginNo);
