+++ src/main/java/com/gongzone/central/member/question/controller/QuestionController.java
@@ -32,21 +32,12 @@ public class QuestionController {
         return statusCode.toString();
     }
 
-    /*@PostMapping("/QuestionStatusUpdate/{memberQuestionNo}")
-    public Mono<String> statusUpdate(@PathVariable int memberQuestionNo, @RequestBody QuestionMember questionMember) {
-        System.out.println("22222222222222222222222222222222222222222");
-        StatusCode statusCode = StatusCode.fromCode(questionMember.getStatusCode());
-        return questionService.getQuestionStatusUpdate(memberQuestionNo, statusCode)
-                .then(Mono.just(statusCode.toString()));
-    }*/
-
     @PostMapping("/QuestionMember/insert")
     public ResponseEntity<Boolean> questionMemberInsert(@RequestBody QuestionMember questionMember) {
         try {
             questionService.getQuestionMemberInsert(questionMember);
             return ResponseEntity.ok(true);
         } catch (Exception e) {
-            System.out.println("실패 : " + e.getMessage());
             return ResponseEntity.ok(false);
         }
     }
