+++ src/main/java/com/gongzone/central/member/question/service/QuestionServiceImpl.java
@@ -31,39 +31,8 @@ public class QuestionServiceImpl implements QuestionService {
         }
     }
 
-    /*@Override
-    public Mono<Void> getQuestionStatusUpdate(int memberQuestionNo, StatusCode statusCode) {
-        if (statusCode == StatusCode.S010701 || statusCode == StatusCode.S010702 || statusCode == StatusCode.S010703) {
-            return Mono.fromRunnable(() -> questionMapper.updateQuestionStatus(memberQuestionNo, statusCode))
-                    .subscribeOn(Schedulers.boundedElastic())
-                    .then(sendAlert(memberQuestionNo));
-        } else {
-            return Mono.empty();
-        }
-    }*/
-
     @Override
     public void getQuestionMemberInsert(QuestionMember questionMember) {
         questionMapper.insertQuestion(questionMember);
     }
-
-    /*private Mono<Void> sendAlert(int memberQuestionNo) {
-        return Mono.fromCallable(() -> {
-                    String memberNo = questionMapper.getMemberNoByQuestionNo(memberQuestionNo);
-
-                    AlertSSE alertSSE = new AlertSSE();
-                    alertSSE.setMemberNo(memberNo);
-                    alertSSE.setTypeCode("T010204"); // 알림 유형 코드
-                    alertSSE.setAlertDetail("테스트.");
-                    return alertSSE;
-                })
-                .flatMap(alertSSE -> {
-                    return alertSEEService.sendAlert(alertSSE);
-                })
-                .subscribeOn(Schedulers.boundedElastic())
-                .doOnError(e -> {
-                    e.printStackTrace();
-                })
-                .doOnSuccess(aVoid -> System.out.println("sendAlert completed successfully"));
-    }*/
 }
