+++ src/main/java/com/gongzone/central/member/report/controller/ReportController.java
@@ -27,27 +27,19 @@ public class ReportController {
 
     @GetMapping("/ReportMember/listReportAll")
     public List<ReportMember> listReportALl() {
-        System.out.println("1111111111111111111111111");
         List<ReportMember> reportMember = reportService.getAllReportMember();
-        System.out.println("reportMember : " + reportMember.size());
         return reportMember;
     }
 
     @PostMapping("/ReportstatusUpdate/{memberNo}")
     public String statusUpdate(@PathVariable String memberNo, @RequestBody RequestReportMember requestReportMember) {
-        System.out.println("11111111111111111111111111");
         StatusCode statusCode = StatusCode.fromCode(requestReportMember.getStatusCode());
-        System.out.println("statusCode : " + statusCode);
-        System.out.println("reportMember.getMemberReportNo : " + requestReportMember.getMemberReportNo());
         reportService.getReportStatusUpdate(requestReportMember.getMemberReportNo(), statusCode);
-        System.out.println("statusCode : " + statusCode);
-        System.out.println("reportMember : " + requestReportMember.toString());
         return statusCode.toString();
     }
 
     @PostMapping("/ReportMember/punish/{memberNo}")
     public ResponseEntity<Boolean> punishInsert(@PathVariable String memberNo, @RequestBody ReportMember reportMember) {
-        System.out.println("11111111111111111111111111");
         MemberPunish memberPunish = MemberPunish.builder()
                 .memberNo(memberNo)
                 .memberAdminNo(reportMember.getMemberAdminNo())
@@ -60,35 +52,25 @@ public class ReportController {
 
         try {
             managementService.getPunishInsert(memberPunish);
-            System.out.println("memberPunish  : " + memberPunish);
 
             StatusCode memberStatusCode = StatusCode.fromCode(reportMember.getMemberStatusCode());
             reportService.getStatusUpdate(memberNo, memberStatusCode);
-            System.out.println("memberStatusCode : " + memberStatusCode);
-            System.out.println("memberNo : " + memberNo);
 
             StatusCode reportStatusCode = StatusCode.fromCode(reportMember.getStatusCode());
             reportService.getReportStatusUpdate(reportMember.getMemberReportNo(), reportStatusCode);
-            System.out.println("reportStatusCode : " + reportStatusCode);
-            System.out.println("reportMember.getMemberReportNo : " + reportMember.getMemberReportNo());
-
 
             return ResponseEntity.ok(true);
         } catch (Exception e) {
-            System.out.println("실패 : " + e.getMessage());
             return ResponseEntity.ok(false);
         }
     }
 
     @PostMapping("/ReportMember/report")
     public ResponseEntity<Boolean> reportMemberInsert(@RequestBody ReportMember reportMember) {
-        System.out.println("11111111111111111111111111");
-        System.out.println("reportMember" + reportMember);
         try {
             reportService.getReportMemberInsert(reportMember);
             return ResponseEntity.ok(true);
         } catch (Exception e) {
-            System.out.println("실패 : " + e.getMessage());
             return ResponseEntity.ok(false);
         }
     }
