+++ src/main/java/com/gongzone/central/member/socialLogin/domain/SocialRequest.java
@@ -1,4 +1,4 @@
-package com.gongzone.central.member.socialLogin.google.domain;
+package com.gongzone.central.member.socialLogin.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class GoogleRequest {
+public class SocialRequest {
     private String code;
     private String state;
     private String userAgent;
