+++ src/main/java/com/gongzone/central/member/socialLogin/naver/controller/NaverController.java
@@ -1,16 +1,17 @@
 package com.gongzone.central.member.socialLogin.naver.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.gongzone.central.member.login.domain.LoginLog;
-import com.gongzone.central.member.login.service.LoginLogService;
-import com.gongzone.central.member.socialLogin.naver.domain.NaverRequest;
-import com.gongzone.central.member.socialLogin.naver.domain.SocialMember;
+import com.gongzone.central.member.socialLogin.domain.SocialMember;
+import com.gongzone.central.member.socialLogin.domain.SocialRequest;
 import com.gongzone.central.member.socialLogin.naver.service.NaverService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 
 @RestController
 @RequestMapping("/api/naver")
@@ -29,7 +30,7 @@ public class NaverController {
             }
 
             String userAgent = request.getHeader("User-Agent");
-            NaverRequest naverRequest = objectMapper.readValue(requestBody, NaverRequest.class);
+            SocialRequest naverRequest = objectMapper.readValue(requestBody, SocialRequest.class);
             naverRequest.setUserAgent(userAgent);
 
             if (naverRequest.getCode() == null || naverRequest.getState() == null) {
